Rename files to match good practices
