+++ src/components/heading/hello-world/HelloWorld.tsx
@@ -1,8 +1,8 @@
 import type { ReactElement } from 'react';
 import React from 'react';
 import type { HelloWorldProps } from './HelloWorld.types';
-import { getAbstractProps } from '@/utils/component/get-abstract-props';
-import { className } from '@/utils/component/class-name';
+import { getAbstractProps } from '@/utils/component/getAbstractProps';
+import { className } from '@/utils/component/className';
 import { useTranslation } from 'react-i18next';
 import { globalLocales } from '@/data/locales/global';
 import { LocalesNS } from '@/data/locales/locales-ns';
