+++ src/storybook/argTypes.collections.ts
@@ -1,6 +1,6 @@
 import type { AbstractProps } from '@/types/component/AbstractProps';
 import type { ArgTypes } from '@storybook/react';
-import { id, className, style, role, tabIndex } from './arg-types';
+import { id, className, style, role, tabIndex } from './argTypes';
 
 export const abstractArgTypes: ArgTypes<AbstractProps> = {
 	id,
