+++ src/utils/component/getAbstractProps.ts
